fix: single prompt command
